# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amrutam Telemedicine Backend",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register",
        "operationId": "register_api_v1_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserRegister"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/setup": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Setup Mfa",
        "operationId": "setup_mfa_api_v1_auth_mfa_setup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MFASetupResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/auth/mfa/verify": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify Mfa",
        "operationId": "verify_mfa_api_v1_auth_mfa_verify_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "secret",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Secret"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/mfa/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Mfa",
        "operationId": "login_mfa_api_v1_auth_mfa_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MFAVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/bookings/slots": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Create Slot",
        "operationId": "create_slot_api_v1_bookings_slots_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilitySlotCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilitySlotResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/bookings/consultations": {
      "get": {
        "tags": [
          "bookings"
        ],
        "summary": "List Consultations",
        "operationId": "list_consultations_api_v1_bookings_consultations_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConsultationResponse"
                  },
                  "type": "array",
                  "title": "Response List Consultations Api V1 Bookings Consultations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Book Consultation",
        "operationId": "book_consultation_api_v1_bookings_consultations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsultationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/bookings/consultations/{consultation_id}": {
      "get": {
        "tags": [
          "bookings"
        ],
        "summary": "Get Consultation",
        "operationId": "get_consultation_api_v1_bookings_consultations__consultation_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consultation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consultation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsultationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/consultations/{consultation_id}/prescription": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Add Prescription",
        "operationId": "add_prescription_api_v1_bookings_consultations__consultation_id__prescription_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consultation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consultation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrescriptionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrescriptionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bookings/consultations/{consultation_id}/payment": {
      "post": {
        "tags": [
          "bookings"
        ],
        "summary": "Process Payment",
        "operationId": "process_payment_api_v1_bookings_consultations__consultation_id__payment_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "consultation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Consultation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentInitiate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AvailabilitySlotCreate": {
        "properties": {
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          }
        },
        "type": "object",
        "required": [
          "start_time",
          "end_time"
        ],
        "title": "AvailabilitySlotCreate"
      },
      "AvailabilitySlotResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "title": "Start Time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "title": "End Time"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "doctor_id",
          "start_time",
          "end_time",
          "status"
        ],
        "title": "AvailabilitySlotResponse"
      },
      "ConsultationCreate": {
        "properties": {
          "slot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Slot Id"
          }
        },
        "type": "object",
        "required": [
          "slot_id"
        ],
        "title": "ConsultationCreate"
      },
      "ConsultationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "patient_id": {
            "type": "string",
            "format": "uuid",
            "title": "Patient Id"
          },
          "doctor_id": {
            "type": "string",
            "format": "uuid",
            "title": "Doctor Id"
          },
          "slot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Slot Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "payment_status": {
            "type": "string",
            "title": "Payment Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "patient_id",
          "doctor_id",
          "slot_id",
          "status",
          "payment_status",
          "created_at"
        ],
        "title": "ConsultationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MFASetupResponse": {
        "properties": {
          "secret": {
            "type": "string",
            "title": "Secret"
          },
          "qr_code_url": {
            "type": "string",
            "title": "Qr Code Url"
          }
        },
        "type": "object",
        "required": [
          "secret",
          "qr_code_url"
        ],
        "title": "MFASetupResponse"
      },
      "MFAVerifyRequest": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 6,
            "minLength": 6,
            "pattern": "^\\d{6}$",
            "title": "Code"
          },
          "temp_token": {
            "type": "string",
            "title": "Temp Token"
          }
        },
        "type": "object",
        "required": [
          "code",
          "temp_token"
        ],
        "title": "MFAVerifyRequest"
      },
...
      "PaymentInitiate": {
        "properties": {
          "consultation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Consultation Id"
          },
          "amount": {
            "type": "number",
            "exclusiveMinimum": 0.0,
            "title": "Amount"
          }
        },
        "type": "object",
        "required": [
          "consultation_id",
          "amount"
        ],
        "title": "PaymentInitiate"
      },
      "PaymentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "consultation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Consultation Id"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "transaction_reference": {
            "type": "string",
            "title": "Transaction Reference"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "consultation_id",
          "amount",
          "transaction_reference",
          "status",
          "created_at"
        ],
        "title": "PaymentResponse"
      },
      "PrescriptionCreate": {
        "properties": {
          "consultation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Consultation Id"
          },
          "diagnosis": {
            "type": "string",
            "minLength": 1,
            "title": "Diagnosis"
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/MedicationItem"
            },
            "type": "array",
            "title": "Medications"
          },
          "doctor_signature": {
            "type": "string",
            "minLength": 1,
            "title": "Doctor Signature"
          }
        },
        "type": "object",
        "required": [
          "consultation_id",
          "diagnosis",
          "medications",
          "doctor_signature"
        ],
        "title": "PrescriptionCreate"
      },
      "PrescriptionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "consultation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Consultation Id"
          },
          "diagnosis": {
            "type": "string",
            "title": "Diagnosis"
          },
          "medications": {
            "items": {
              "$ref": "#/components/schemas/MedicationItem"
            },
            "type": "array",
            "title": "Medications"
          },
          "doctor_signature": {
            "type": "string",
            "title": "Doctor Signature"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "consultation_id",
          "diagnosis",
          "medications",
          "doctor_signature",
          "created_at"
        ],
        "title": "PrescriptionResponse"
      },
      "StatusResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "title": "StatusResponse"
      },
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          },
          "is_mfa_required": {
            "type": "boolean",
            "title": "Is Mfa Required",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token"
      },
      "UserLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin"
      },
      "UserRegister": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "title": "Password",
            "description": "Password must be at least 8 characters long"
          },
          "role": {
            "type": "string",
            "title": "Role",
            "description": "Role: patient, doctor, admin",
            "default": "patient"
          },
          "first_name": {
            "type": "string",
            "minLength": 1,
            "title": "First Name"
          },
          "last_name": {
            "type": "string",
            "minLength": 1,
            "title": "Last Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone",
            "description": "Phone number"
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "title": "Date Of Birth"
          },
          "specialty": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Specialty"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "consultation_fee": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Consultation Fee"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password",
          "first_name",
          "last_name",
          "phone",
          "date_of_birth"
        ],
        "title": "UserRegister"
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "is_mfa_enabled": {
            "type": "boolean",
            "title": "Is Mfa Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "role",
          "is_mfa_enabled",
          "created_at"
        ],
        "title": "UserResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}